+++ packages/asc-ui/src/components/Checkbox/index.ts
@@ -1,2 +1,2 @@
-export { default as CheckboxStyle } from './CheckboxStyle'
+export { default as CheckboxStyle, CheckboxWrapperStyle } from './CheckboxStyle'
 export { default as Checkbox } from './Checkbox'