+++ packages/asc-ui/src/components/Label/Label.tsx
@@ -5,6 +5,7 @@ import LabelContext from './LabelContext'
 
 type Props = {
   label: string
+  noActiveState?: boolean // Temporary solution to make the active state on the label optional, as there is nothing specified in design system. Needs to be discussed with design.
 } & StyleProps
 
 const Label: React.FC<Props & React.LabelHTMLAttributes<HTMLLabelElement>> = ({
@@ -12,20 +13,31 @@ const Label: React.FC<Props & React.LabelHTMLAttributes<HTMLLabelElement>> = ({
   label,
   disabled,
   position,
+  noActiveState,
   ...otherProps
 }) => {
   const [active, setActive] = React.useState(false)
   const { children } = usePassPropsToChildren(childrenProps, {
     disabled,
   })
+
+  const activeState = noActiveState ? false : active
+
   return (
     <LabelContext.Provider value={{ active, setActive }}>
-      <LabelStyle {...{ ...otherProps, disabled, position, active }}>
+      <LabelStyle
+        {...{ ...otherProps, disabled, position }}
+        active={activeState}
+      >
         <LabelTextStyle position={position}>{label}</LabelTextStyle>
         {children}
       </LabelStyle>
     </LabelContext.Provider>
   )
 }
+
+Label.defaultProps = {
+  noActiveState: false,
+} as Props
 
 export default Label