+++ CHANGELOG.md
@@ -16,12 +16,16 @@ Prefix the change with one of these keywords:
 
 ## UNRELEASED
 
+- Added: default `position` prop for Label(Text) is 'right'
+- Added: Checkbox has Error state, by passing an `error` prop
+- Fixed: updated Checkbox according to design system
+
 ## [0.17.2] - 2019-10-14
 
 ### ASC-UI
 
 - Added: accessibility support for MenuFlyOut to navigate with arrow keys when open
-- Fixed: firefox issue with blurring on MenuFlyOut 
+- Fixed: firefox issue with blurring on MenuFlyOut
 
 ## [0.17.0] - 2019-10-14
 
@@ -40,7 +44,7 @@ Prefix the change with one of these keywords:
 
 ## [0.15.5] - 2019-09-17
 
-- Added: theme spacing unit and a themeSpacing util to retrieve a size from the theme's spacing unit. 
+- Added: theme spacing unit and a themeSpacing util to retrieve a size from the theme's spacing unit.
 - Fixed: issue with button's size
 - Changed: The `max-width` value of the header is now defined by the max width of the grid in the theme configuration
 - Deprecated: `fromTheme` util method. Use `getValueFromTheme` instead